new: Using `maybe_async` to generate Parser async and sync api
