+++ .github/workflows/test.yaml
@@ -18,7 +18,7 @@ jobs:
     strategy:
       matrix:
         os: [ "ubuntu-latest", 'windows-latest', 'macos-latest' ]
-        features: [ " ", "serde", "tokio" ]
+        features: [ " ", "serde", "async", "sync" ]
 
     steps:
       - name: Disable autocrlf
