+++ crates/cw/Cargo.toml
@@ -40,7 +40,7 @@ features = ["io-util"]
 
 [dependencies.libcw]
 path = "../libcw"
-features = ["tokio","serde"]
+features = ["serde", "async"]
 
 [dependencies.clap]
 version = "3.1.0"
@@ -62,7 +62,6 @@ version = "3.1.0"
 
 [build-dependencies.libcw]
 path = "../libcw"
-features = ["tokio"]
 
 [build-dependencies.clap]
 version = "3.1.0"
